Backport r993951 by lvsouza from trunk to the 4.2 branch:
kppplogview uses int to count bytes in and out in a session and to sum up bytes
in and out during a month, when any of those variables reaches 2 GB an overflow
happens. This patch changes variable types from 32-bit to 64-bit.
CCBUG: 196839


svn path=/branches/KDE/4.2/kdenetwork/kppp/; revision=1002338

+++ logview/monthly.cpp
@@ -35,16 +35,16 @@
 #include "export.h"
 #include <qstringlist.h>
 
-static void formatBytes(int bytes, QString &result) {
+static void formatBytes(long int bytes, QString &result) {
   if(bytes < 1024)
     result.setNum(bytes);
   else if(bytes < 1024*1024)
-    result = i18n("%1 KiB", KGlobal::locale()->formatNumber((float)bytes / 1024.0, 1));
+    result = i18n("%1 KiB", KGlobal::locale()->formatNumber((double)bytes / 1024.0, 1));
   else
-    result = i18n("%1 MiB", KGlobal::locale()->formatNumber((float)bytes / 1024.0 / 1024.0, 1));
+    result = i18n("%1 MiB", KGlobal::locale()->formatNumber((double)bytes / 1024.0 / 1024.0, 1));
 }
 
-static void formatBytesMonth(int bytes, QString &result) {
+static void formatBytesMonth(long int bytes, QString &result) {
 
   int day, days;
   day = QDate::currentDate().day();
@@ -55,9 +55,9 @@ static void formatBytesMonth(int bytes, QString &result) {
   if(bytes < 1024)
     result.setNum(bytes);
   else if(bytes < 1024*1024)
-    result = i18n("%1 KiB", KGlobal::locale()->formatNumber((float)bytes / 1024.0, 1));
+    result = i18n("%1 KiB", KGlobal::locale()->formatNumber((double)bytes / 1024.0, 1));
   else
-    result = i18n("%1 MiB", KGlobal::locale()->formatNumber((float)bytes / 1024.0 / 1024.0, 1));
+    result = i18n("%1 MiB", KGlobal::locale()->formatNumber((double)bytes / 1024.0 / 1024.0, 1));
 }
 
 static void formatDuration(int seconds, QString &result) {
@@ -165,10 +165,10 @@ QString LogListItem::key(int c, bool ascending) const
     k.sprintf("%012.2f", li->sessionCosts());
     break;
   case 6:
-    k.sprintf("%012d", li->bytesIn());
+    k.sprintf("%012ld", li->bytesIn());
     break;
   case 7:
-    k.sprintf("%012d", li->bytesOut());
+    k.sprintf("%012ld", li->bytesOut());
     break;
   }
   return k;
@@ -291,7 +291,7 @@ void MonthlyWidget::plotMonth() {
   // for collecting monthly statistics
   int count = 0;
   double costs = 0;
-  int bin = 0, bout = 0;
+  long int bin = 0, bout = 0;
   int duration = 0;
   lv->clear();
   selectionItem = 0L;
@@ -506,7 +506,7 @@ void MonthlyWidget::exportWizard() {
   // for collecting monthly statistics
   int count = 0;
   double costs = 0;
-  int bin = 0, bout = 0;
+  long int bin = 0, bout = 0;
   int duration = 0;
 
   for(int i = 0; i < (int)logList.count(); i++) {
@@ -667,7 +667,7 @@ void MonthlyWidget::slotSelectionChanged()
   {
     int count = 0;
     double costs = 0;
-    int bin = 0, bout = 0;
+    long int bin = 0, bout = 0;
     int duration = 0;
     LogListItem *item;
     LogInfo *li;
